requirements.txt
pandas>=2.3.3
numpy>=2.3.0
requests>=2.32.3
yfinance>=0.2.44
PyYAML>=6.0.2
APScheduler>=3.10.4
fastapi==0.115.0
uvicorn[standard]==0.30.6
Jinja2==3.1.4
